… performance/init.sql
… performance/initial_analysis.md