### Added Indexes
- `catalog_item(catalog_id)`
- `shopping_list(user_id)`
- `catalog_item(food_id, price)`

### Performance Improvements
- **Store Catalog Query**
  - **Endpoint:** `/stores/{store_id}/catalog`
  - **Execution Time:** 8.389ms → 0.972ms (88% improvement)
  
- **User Shopping Lists Query**
  - **Endpoint:** `/users/{user_id}/lists`
  - **Execution Time:** 10.218ms → 0.030ms (99.7% improvement)
  
- **Route Optimization Query**
  - **Endpoint:** `/shopping/route_optimize`
  - **Execution Time:** 11.536ms → 1.953ms (83% improvement)

### Other Changes
- Removed `performance/initial_analysis.md`
- Updated `schema.sql` with new index definitions